+++ src/index.ts
@@ -1,5 +1,5 @@
 export { generate, validateSpec } from "./core/generator";
-export { parseOpenAPI } from "./core/parser";
+export { parseOpenAPI, convertSwagger2ToOpenApi3, isSwagger2Document } from "./core/parser";
 export { extractHandlers, injectHandlers, isCustomFile, CUSTOM_FILES, GENERATED_REGION_START, GENERATED_REGION_END } from "./core/incremental";
 export { filterTools, groupTools, parseTagList, parsePathList, parseGroupBy, loadOperationAllowlistFile, resolveAllowlistValue, matchPathPattern, matchesAnyPattern, globToRegExp, toGroupMetadata } from "./core/filter-group";
 export { listKnownSpecs, getSpecInfo, fetchSpecToCwd, KNOWN_SPECS } from "./core/registry";